Config defaults confirmation. Untracked unneeded script. Added support for native menu on tray items.

+++ Bar/ShutdownTimerWidget.qml
@@ -12,11 +12,13 @@ Item {
     id: button
 
     text: {
-      if (shutdown.timeToTargetTimeSeconds <= 900 && Config.parsedConfig.miscellaneous.shutdownWidget.enableShutdown.value) {
-        if (shutdown.timeToTargetTimeSeconds <= shutdown.responseTime) {
-          shutdown.timeToTargetTime + "!"
-        } else {
-          shutdown.timeToTargetTime
+      if (Config.parsedConfig.miscellaneous.shutdownWidget) {
+        if (shutdown.timeToTargetTimeSeconds <= 900 && Config.parsedConfig.miscellaneous.shutdownWidget.enableShutdown.value) {
+          if (shutdown.timeToTargetTimeSeconds <= shutdown.responseTime) {
+            shutdown.timeToTargetTime + "!"
+          } else {
+            shutdown.timeToTargetTime
+          }
         }
       }
     }
@@ -35,8 +37,10 @@ Item {
       running: true
 
       onTriggered: {
-        if (button.shutdown.timeToTargetTimeSeconds < 0 && Config.parsedConfig.miscellaneous.shutdownWidget.enableShutdown.value) {
-          Quickshell.execDetached(["systemctl", "poweroff"])
+        if (Config.parsedConfig.miscellaneous.shutdownWidget) {
+          if (button.shutdown.timeToTargetTimeSeconds < 0 && Config.parsedConfig.miscellaneous.shutdownWidget.enableShutdown.value) {
+            Quickshell.execDetached(["systemctl", "poweroff"])
+          }
         }
       }
     }
